Add move dataHelper into dataHelper directory, create test file
